Refactor to use its own command
